+++ common/src/commonMain/kotlin/dev/inmo/micro_utils/common/ToFixed.kt
@@ -1,0 +1,6 @@
+package dev.inmo.micro_utils.common
+
+val FixedSignsRange = 0 .. 100
+
+expect fun Float.fixed(signs: Int): Float
+expect fun Double.fixed(signs: Int): Double