+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.17.3
 
 * `Common`:
+    * Add `fixed` extensions for `Float` and `Double`
     * New function `emptyDiff`
     * Now you may pass custom `comparisonFun` to all `diff` functions