+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.17.3
+
+* `Common`:
+    * Add `fixed` extensions for `Float` and `Double`
+    * New function `emptyDiff`
+    * Now you may pass custom `comparisonFun` to all `diff` functions
+
 ## 0.17.2
 
 * `FSM`: